bluetooth: backport to 2.6.29
[ARM] msm: msm_serial_hs: Rename wakeup to rx_wakeup.

Preparation for a new low power wakeup mode.

Signed-off-by: Nick Pelly <npelly@google.com>

[ARM] msm: msm_serial_hs: Improve low power mode support.

Move the request_clock_off()/on() into the header.
Add exit_lpm_cb callback to help exit low power mode.
Clean up RTS / RFR handling. Deactivate RFR when clocking off the uart
unless RX wakeup lpm is being used.

Signed-off-by: Nick Pelly <npelly@google.com>

[ARM] msm: msm_serial_hs: Wait for RX fifo to empty before clocking off.

This is needed if we clock off quickly after RX.

Signed-off-by: Nick Pelly <npelly@google.com>

[ARM] msm: msm_serial_hs: Do not allow suspend while BT DMA could be active.

Signed-off-by: Nick Pelly <npelly@google.com>

[ARM] msm: msm_serial_hs: Improve RX flush procedure during clock off sequence.

Clock off is requested by msm_hs_request_clock_off. The new procedure is:
1) Immediately deassert RTS to prevent new RX data.
2) Wait for the TX pipeline to flush (TXLEV, TXEMT).
3) Issue a FORCE_STALE_EVENT command, and flush on RXSTALE.
5) After this flush completes, disable DLINK and flush one more time
   (discarding data).
6) Finally clock off the uart.

The key changes are to disable DLINK - which fixes the datamover lockup, and
the issuing of the FORCE_STALE_EVENT command - which prevents RX data loss
during clock off sequence.

Signed-off-by: Nick Pelly <npelly@google.com>

[ARM] msm: Add 'execute' datamover callback

Based on a patch from Brent DeGraaf:

"The datamover supports channels which can be shared amongst devices.
As a result, the actual data transfer may occur some time after the
request is queued up. Some devices such as mmc host controllers
will timeout if a command is issued too far in advance of the actual
transfer, so if dma to other devices on the same channel is already
in progress or queued up, the added delay can cause pending transfers
to fail before they start. This change extends the api to allow a
user callback to be invoked just before the actual transfer takes
place, thus allowing actions directly associated with the dma
transfer, such as device commands, to be invoked with precise timing.
Without this mechanism, there is no way for a driver to realize
this timing. Also adds a user pointer to the command structure for use
by the caller to reference information that may be needed by the
callback routine for proper identification and processing associated
with that specific request. This change is necessary to fix problems
associated with excessive command timeouts and race conditions in the
mmc driver."

This patch also fixes all the callers of msm_dmov_enqueue_cmd() to
ensure their callback function is NULL.

Signed-off-by: San Mehat <san@google.com>
Cc: Brent DeGraaf <bdegraaf@quicinc.com>
Cc: Brian Swetland <swetland@google.com>

msm_serial_hs: Use dedicated workqueue thread for tty RX.

Previously the tty layer used the global workqueue, which can introduce high
latency on the RX path if other global workqueue users block.

Introduce a dedicated msm_serial_hs workqueue thread - managed by the
msm_serial_hs drvier - to call tty_flip_buffer_push(). Set tty low_latency
to 1 so that the tty layer does not use its own work queue.

This patch fixes an issue where A2DP would skip when the screen turns on.

Change-Id: I28253f8e5b0f577c7fc635ff5b10cf2abc042245
Signed-off-by: Nick Pelly <npelly@google.com>

Bluetooth: Add ACL MTU, available buffers and total buffers to hci_conn_info.

This provides userspace debugging tools access to ACL flow control state.

Signed-off-by: Nick Pelly <npelly@google.com>

Bluetooth: Introduce L2CAP_LM_FLUSHABLE to allow flushing of ACL packets.

With Bluetooth 2.1 ACL packets can be flushable or non-flushable. This changes
makes the default ACL packet non-flushable, and allows selection of flushable
packets on a per-L2CAP socket basis with L2CAP_LM_FLUSHABLE.

Note the HCI Write Automatic Flush Timeout command also needs to be issued
to set the flush timeout to non-zero.

Need to featurize this change to Bluetooth 2.1 chipsets only before pushing
upstream.

Signed-off-by: Nick Pelly <npelly@google.com>

Bluetooth: Hack: Do not use power_save feature.

power_save is a feature to allow HID devices to control the sniff mode.
Unfortunately it also prevents us exiting sniff mode on some A2DP devices
that do not explicitly exit sniff mode themselves, resulting in skipping
audio.

Marcel is trying to finalize a setsockopt() API to control whether power_save
is enabled on a per socket basis. In the mean-time, turn off power_save for
Android, since we do not have official HID support, and this is causing
problems for A2DP which we do support.

(Note the power_save logic is reversed, 0 is on, 1 is off).

Change-Id: Ife4478055128b81669bf49308d2e2199e1aa11a1
Signed-off-by: Nick Pelly <npelly@google.com>

Bluetooth: Allow SCO/eSCO packet type selection for outgoing SCO connections.

__u16 sco_pkt_type is introduced to struct sockaddr_sco. It allows bitwise
selection of SCO/eSCO packet types. Currently those bits are:

0x0001 HV1 may be used.
0x0002 HV2 may be used.
0x0004 HV3 may be used.
0x0008 EV3 may be used.
0x0010 EV4 may be used.
0x0020 EV5 may be used.
0x0040 2-EV3 may be used.
0x0080 3-EV3 may be used.
0x0100 2-EV5 may be used.
0x0200 3-EV5 may be used.

This is similar to the Packet Type parameter in the HCI Setup Synchronous
Connection Command, except that we are not reversing the logic on the EDR bits.
This makes the use of sco_pkt_tpye forward portable for the use case of
white-listing packet types, which we expect will be the primary use case.

If sco_pkt_type is zero, or userspace uses the old struct sockaddr_sco,
then the default behavior is to allow all packet types.

Packet type selection is just a request made to the Bluetooth chipset, and
it is up to the link manager on the chipset to negiotiate and decide on the
actual packet types used. Furthermore, when a SCO/eSCO connection is eventually
made there is no way for the host stack to determine which packet type was used
(however it is possible to get the link type of SCO or eSCO).

sco_pkt_type is ignored for incoming SCO connections. It is possible
to add this in the future as a parameter to the Accept Synchronous Connection
Command, however its a little trickier because the kernel does not
currently preserve sockaddr_sco data between userspace calls to accept().

The most common use for sco_pkt_type will be to white-list only SCO packets,
which can be done with the hci.h constant SCO_ESCO_MASK.

This patch is motivated by broken Bluetooth carkits such as the Motorolo
HF850 (it claims to support eSCO, but will actually reject eSCO connections
after 5 seconds) and the 2007/2008 Infiniti G35/37 (fails to route audio
if a 2-EV5 packet type is negiotiated). With this patch userspace can maintain
a list of compatible packet types to workaround remote devices such as these.

Based on a patch by Marcel Holtmann.

Change-Id: I304d8fda5b4145254820a3003820163bf53de5a5
Signed-off-by: Nick Pelly <npelly@google.com>

[ARM] msm: clock: hack: Always set axi clock to max rate when enabling uartdm.

This implementation is a gross hack. We are working with Qualcomm to understand
this further and get a better fix (perhaps in the radio).

Signed-off-by: Nick Pelly <npelly@google.com>

bluetooth: compilation fixes